modified the lapic_id type to uint8_t
According intel mannual and ACPI mannual,lapic_id length is 1 byte.

V1->V2:
  Add U suffix to the numeric when do arithmetic operation on lapic.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/arch/x86/lapic.h
@@ -160,7 +160,7 @@ void save_lapic(struct lapic_regs *regs);
 int early_init_lapic(void);
 int init_lapic(uint16_t cpu_id);
 void send_lapic_eoi(void);
-uint32_t get_cur_lapic_id(void);
+uint8_t get_cur_lapic_id(void);
 int send_startup_ipi(enum intr_cpu_startup_shorthand cpu_startup_shorthand,
 		uint32_t cpu_startup_dest,
 		uint64_t cpu_startup_start_address);
